refactored folder structure of ui elements
