+++ src/code-generator/entry-points/ui-elements/index.ts
@@ -1,0 +1,3 @@
+export * from "./spinner";
+export { spinnerTexts as interactiveCliTexts } from "./interactive-cli";
+export { spinnerTexts as nonInteractiveCliTexts } from "./non-interactive-cli";
