+++ src/code-generator/entry-points/ui-elements/interactive-cli/index.ts
@@ -1,0 +1,1 @@
+export const spinnerTexts = {};
