+++ src/code-generator/entry-points/ui-elements/non-interactive-cli/index.ts
@@ -1,0 +1,9 @@
+export const spinnerTexts = {
+  onStart: "Generating an unmatched application starter",
+  onSucceed:
+    "All set, the code is ready for you with tons of good practices inside\nDeepen your understanding by reading our article 'Coding with Practica': https://practica.dev/the-basics/coding-with-practica",
+  onError: {
+    default:
+      "Embarrassingly our code generator failed. Yeah, almost 100% test coverage did not help here. Could you be nice to us and open an issue?",
+  },
+};
