+++ src/code-generator/entry-points/ui-elements/spinner/index.ts
