--- src/code-generator/utils/cli-texts.ts
@@ -1,11 +1,0 @@
-export default {
-  nonInteractiveCli: {
-    onStart: "Generating an unmatched application starter",
-    onSucceed:
-      "All set, the code is ready for you with tons of good practices inside\nDeepen your understanding by reading our article 'Coding with Practica': https://practica.dev/the-basics/coding-with-practica",
-    onError: {
-      default:
-        "Embarrassingly our code generator failed. Yeah, almost 100% test coverage did not help here. Could you be nice to us and open an issue?",
-    },
-  },
-};
