--- src/code-generator/utils/index.ts
@@ -1,3 +1,0 @@
-import cliTexts from "./cli-texts";
-export * from "./spinner";
-export { cliTexts };
